preprocess unify and disunify
