+++ packages/backend/helpers/crm/transform/unify.ts
@@ -1,7 +1,7 @@
 import { TP_ID, accountFieldMappingConfig } from '@prisma/client';
 import { StandardObjects } from '../../../constants/common';
-import { PipedriveDealStatus } from '../../../constants/pipedrive';
 import { transformFieldMappingToModel } from '.';
+import { preprocessUnifyObject } from './preprocess';
 
 export async function unifyObject<T extends Record<string, any>, K>({
     obj,
@@ -16,7 +16,7 @@ export async function unifyObject<T extends Record<string, any>, K>({
     tenantSchemaMappingId?: string;
     accountFieldMappingConfig?: accountFieldMappingConfig;
 }): Promise<K> {
-    const processedObj = preprocessObject({ obj, tpId, objType });
+    const processedObj = preprocessUnifyObject({ obj, tpId, objType });
     const transformedObject = await transformFieldMappingToModel({
         obj: processedObj,
         tpId,
@@ -40,34 +40,3 @@ export async function unifyObject<T extends Record<string, any>, K>({
 
     return unifiedObject as K;
 }
-
-const preprocessObject = <T extends Record<string, any>>({
-    obj,
-    tpId,
-    objType,
-}: {
-    obj: T;
-    tpId: TP_ID;
-    objType: StandardObjects;
-}) => {
-    const preprocessMap: any = {
-        [TP_ID.pipedrive]: {
-            [StandardObjects.deal]: (obj: T) => {
-                return {
-                    ...obj,
-                    revert_isWon: obj.status === PipedriveDealStatus.won,
-                };
-            },
-        },
-        [TP_ID.zohocrm]: {
-            [StandardObjects.deal]: (obj: T) => {
-                return {
-                    ...obj,
-                    revert_isWon: obj.Stage === 'Closed (Won)',
-                };
-            },
-        },
-    };
-    const transformFn = (preprocessMap[tpId] || {})[objType];
-    return transformFn ? transformFn(obj) : obj;
-};
